+++ libEM/ctf.cpp
@@ -299,7 +299,7 @@ void SimpleCtf::compute_2d_complex(EMData * image, CtfType type, XYData * sf)
 			int ynx = y * nx;
 
 			for (int x = 0; x < nx / 2; x++) {
-				float s = hypot(x, y - ny / 2.0f) * ds;
+				float s = (float) hypot(x, y - ny / 2.0f) * ds;
 				d[x * 2 + ynx] = calc_noise(s);
 				d[x * 2 + ynx + 1] = d[x * 2 + ynx];
 			}
@@ -310,7 +310,7 @@ void SimpleCtf::compute_2d_complex(EMData * image, CtfType type, XYData * sf)
 			int ynx = y * nx;
 
 			for (int x = 0; x < nx / 2; x++) {
-				float s = hypot(x, y - ny / 2.0f) * ds;
+				float s = (float)hypot(x, y - ny / 2.0f) * ds;
 				float gamma = calc_gamma(g1, g2, s);
 				float v = calc_amplitude(gamma);
 				d[x * 2 + ynx] *= v;
@@ -323,7 +323,7 @@ void SimpleCtf::compute_2d_complex(EMData * image, CtfType type, XYData * sf)
 			int ynx = y * nx;
 
 			for (int x = 0; x < nx / 2; x++) {
-				float s = hypot((float) x, (float) y - ny / 2) * ds;
+				float s = (float)hypot((float) x, (float) y - ny / 2) * ds;
 				float gamma = calc_gamma(g1, g2, s);
 				float v = fabs(calc_amplitude(gamma));
 				d[x * 2 + ynx] *= v;
@@ -335,7 +335,7 @@ void SimpleCtf::compute_2d_complex(EMData * image, CtfType type, XYData * sf)
 		for (int y = 0; y < ny; y++) {
 			int ynx = y * nx;
 			for (int x = 0; x < nx / 2; x++) {
-				float s = hypot(x, y - ny / 2.0f) * ds;
+				float s = (float)hypot(x, y - ny / 2.0f) * ds;
 				float gamma = calc_gamma(g1, g2, s);
 				float v = calc_amplitude(gamma);
 				v = v > 0 ? 1.0f : -1.0f;
@@ -350,7 +350,7 @@ void SimpleCtf::compute_2d_complex(EMData * image, CtfType type, XYData * sf)
 			int ynx = y * nx;
 
 			for (int x = 0; x < nx / 2; x++) {
-				float s = hypot(x, y - ny / 2.0f) * ds;
+				float s = (float)hypot(x, y - ny / 2.0f) * ds;
 				float v = exp(-(bfactor * s * s));
 				d[x * 2 + ynx] *= v;
 				d[x * 2 + ynx + 1] = d[x * 2 + ynx];
@@ -365,7 +365,7 @@ void SimpleCtf::compute_2d_complex(EMData * image, CtfType type, XYData * sf)
 
 			for (int x = 0; x < nx / 2; x++) {
 
-				float s = hypot(x, y - ny / 2.0f) * ds;
+				float s = (float)hypot(x, y - ny / 2.0f) * ds;
 				float gamma = calc_gamma(g1, g2, s);
 				float f = calc_ctf1(amp1, gamma, s);
 				float noise = calc_noise(s);
